+++ z0lib/z0librun.py
@@ -25,9 +25,15 @@ import configparser
 import inspect
 import sys
 import os
+import os.path as pth
+from os.path import expanduser as pthuser
+from datetime import datetime
+import re
 from builtins import object
 import shutil
 import shlex
+
+
 if sys.version_info[0] == 2:  # pragma: no cover
     from subprocess import PIPE, Popen
     DEVNULL = open("/dev/null", "w").fileno()
@@ -52,11 +58,17 @@ ODOO_CONF = [
 # Read Odoo configuration file (False or /etc/openerp-server.conf)
 OE_CONF = False
 DEFDCT = {}
-__version__ = "2.0.17"
+RMODE = "rU" if sys.version_info[0] == 2 else "r"
+
+__version__ = "2.1.0"
+
+
+def os_env(key, default=None):
+    return os.environ.get(key, default)
 
 
 def nakedname(path):
-    return os.path.splitext(os.path.basename(path))[0]
+    return pth.splitext(pth.basename(path))[0]
 
 
 def print_flush(msg, end=None, flush=True):
@@ -68,19 +80,32 @@ def print_flush(msg, end=None, flush=True):
             sys.stdout.flush()
 
 
-def echo_cmd_verbose(args, dry_run=False, os_level=0, flush=False):  # pragma: no cover
+def echo_cmd_verbose(
+        args, dry_run=False, os_level=0, humdrum=1, flush=False):  # pragma: no cover
     prompt = ">" if dry_run else "$"
     prompt = ("  " * os_level) + prompt
+    eol = ""
+    if humdrum == 0 and os_env("PS_RUN_COLOR"):
+        prompt += "\033[" + os.environ["PS_RUN_COLOR"] + "m"
+        if os_env("PS_NOP_COLOR"):
+            eol = "\033[" + os.environ["PS_NOP_COLOR"] + "m"
+        else:
+            eol = "\033[0m"
     if isinstance(args, (tuple, list)):
         if flush and sys.version_info[0] == 3:
-            print_flush("%s %s" % (prompt, join_args(args)), flush=flush)
+            print_flush("%s %s%s" % (prompt, join_args(args), eol), flush=flush)
         else:
-            print_flush("%s %s" % (prompt, join_args(args)))
+            print_flush("%s %s%s" % (prompt, join_args(args), eol))
     else:
         if flush and sys.version_info[0] == 3:
-            print_flush("%s %s" % (prompt, args), flush=flush)
+            print_flush("%s %s%s" % (prompt, args, eol), flush=flush)
         else:
-            print_flush("%s %s" % (prompt, args))
+            print_flush("%s %s%s" % (prompt, args, eol))
+
+
+def clear_tty(dry_run=None, flush=False):
+    if not dry_run:
+        print_flush("\033[0m", end="", flush=flush)
 
 
 def join_args(args):
@@ -104,7 +129,14 @@ def join_args(args):
 
 
 def os_system_traced(
-        args, verbose=0, dry_run=None, with_shell=None, rtime=None, os_level=0):
+        args,
+        verbose=0,
+        dry_run=None,
+        with_shell=None,
+        rtime=None,
+        os_level=0,
+        flush=None,
+        humdrum=1):
     # Execute <os.system> like function and return sts, stdout, stderr
 
     def read_from_proc_n_echo(proc, outerr, rtime):
@@ -126,13 +158,7 @@ def os_system_traced(
 
     joined_args = join_args(args) if isinstance(args, (tuple, list)) else args
     if verbose:
-        # prompt = ">" if dry_run else "$"
-        # prompt = ("  " * os_level) + prompt
-        # if isinstance(args, (tuple, list)):
-        #     print('%s %s' % (prompt, joined_args))
-        # else:
-        #     print('%s %s' % (prompt, args))
-        echo_cmd_verbose(args, dry_run=dry_run, os_level=os_level)
+        echo_cmd_verbose(args, dry_run=dry_run, os_level=os_level, humdrum=humdrum)
 
     rtime = True if rtime is None else rtime
     with_shell = True if with_shell is None else with_shell
@@ -152,11 +178,11 @@ def os_system_traced(
             sts = proc.wait()
         except OSError as e:  # noqa: F821
             if verbose:
-                print_flush(e, flush=True)
+                print_flush(e, flush=flush)
             sts = 127
         except BaseException as e:
             if verbose:
-                print_flush(e, flush=True)
+                print_flush(e, flush=flush)
             sts = 126
     else:
         try:
@@ -172,22 +198,31 @@ def os_system_traced(
                 sts = proc.wait()
         except FileNotFoundError as e:  # noqa: F821
             if verbose:
-                print_flush(e, flush=True)
+                print_flush(e, flush=flush)
             sts = 127
         except BaseException as e:
             if verbose:
-                print_flush(e, flush=True)
+                print_flush(e, flush=flush)
             sts = 126
+    clear_tty(dry_run=dry_run, flush=flush)
     return sts, prcout, prcerr
 
 
 def os_system(
-        args, verbose=0, dry_run=None, with_shell=True, rtime=True, os_level=0):
+        args,
+        verbose=0,
+        dry_run=None,
+        with_shell=True,
+        rtime=True,
+        os_level=0,
+        flush=None,
+        humdrum=1):
     # Execute <os.system> like function (return just sts)
 
     joined_args = join_args(args) if isinstance(args, (tuple, list)) else args
     if verbose:
-        echo_cmd_verbose(args, dry_run=dry_run, os_level=os_level)
+        echo_cmd_verbose(
+            args, dry_run=dry_run, os_level=os_level, humdrum=humdrum, flush=flush)
 
     rtime = True if rtime is None else rtime
     with_shell = True if with_shell is None else with_shell
@@ -204,11 +239,11 @@ def os_system(
             sts = proc.wait()
         except OSError as e:  # noqa: F821
             if verbose:
-                print_flush(e)
+                print_flush(e, flush=flush)
             sts = 127
         except BaseException as e:
             if verbose:
-                print_flush(e)
+                print_flush(e, flush=flush)
             sts = 126
     else:
         try:
@@ -223,12 +258,13 @@ def os_system(
                 sts = proc.wait()
         except FileNotFoundError as e:  # noqa: F821
             if verbose:
-                print_flush(e)
+                print_flush(e, flush=flush)
             sts = 127
         except BaseException as e:
             if verbose:
-                print_flush(e)
+                print_flush(e, flush=flush)
             sts = 126
+    clear_tty(dry_run=dry_run, flush=flush)
     return sts
 
 
@@ -237,7 +273,9 @@ def run_traced(cmd,
                dry_run=None,
                disable_alias=None,
                is_alias=None,
-               rtime=False):
+               rtime=False,
+               flush=True,
+               humdrum=0):
     """Run system command with aliases log; return system status and trace log"""
 
     def cmd_neutral_if(
@@ -316,41 +354,37 @@ def run_traced(cmd,
                 paths.append(arg)
         return argv, opt_unk, paths, params
 
-    def sh_any(args, verbose=0, dry_run=None):
-        # prcout = prcerr = ""
-        # if (
-        #     cmd_neutral_if(args, params=['dir'])
-        #     or cmd_neutral_if(args, params=['ls'])
-        # ):
-        #     dry_run = False
-        # if dry_run:
-        #     return 0, prcout, prcerr
+    def sh_any(args, verbose=0, dry_run=None, flush=None, humdrum=1):
         with_shell = False
         while 1:
             sts, prcout, prcerr = os_system_traced(
                 args, verbose=verbose, dry_run=dry_run,
-                with_shell=with_shell, rtime=rtime)
+                with_shell=with_shell, rtime=rtime,
+                flush=flush, humdrum=humdrum)
             if sts == 0 or with_shell or rtime:
                 break
             with_shell = True
+        clear_tty(dry_run=dry_run, flush=flush)
         return sts, prcout, prcerr
 
-    def sh_cd(args, verbose=0, dry_run=None):
+    def sh_cd(args, verbose=0, dry_run=None, flush=None, humdrum=1):
         if verbose:
-            echo_cmd_verbose(args, dry_run=dry_run)
+            echo_cmd_verbose(args, dry_run=dry_run, humdrum=humdrum)
         argv, opt_unk, paths, params = simple_parse(args, {})
         sts = 0
         tgtpath = paths[0] if paths else os.environ["HOME"]
-        if os.path.isdir(tgtpath):
+        if pth.isdir(tgtpath):
             os.chdir(tgtpath)
         elif not dry_run:
             sts = 1
+        clear_tty(dry_run=dry_run, flush=flush)
         return sts, "", ""
 
-    def sh_cp(args, verbose=0, dry_run=None):
+    def sh_cp(args, verbose=0, dry_run=None, flush=None, humdrum=1):
         if dry_run:
             if verbose:
-                echo_cmd_verbose(args, dry_run=dry_run)
+                echo_cmd_verbose(args, dry_run=dry_run, humdrum=humdrum)
+            clear_tty(dry_run=dry_run, flush=flush)
             return 0, "", ""
         argv, opt_unk, paths, params = simple_parse(
             args, {
@@ -363,12 +397,12 @@ def run_traced(cmd,
         )
         if not opt_unk and paths[0] and paths[1]:
             if verbose:
-                echo_cmd_verbose(args, dry_run=dry_run)
+                echo_cmd_verbose(args, dry_run=dry_run, humdrum=humdrum)
             if (
                 paths[1]
-                and os.path.basename(paths[1]) != os.path.basename(paths[0])
+                and pth.basename(paths[1]) != pth.basename(paths[0])
             ):
-                paths[1] = os.path.join(paths[1], os.path.basename(paths[0]))
+                paths[1] = pth.join(paths[1], pth.basename(paths[0]))
             if params["-r"] or params["-R"] or params["--recursive"]:
                 if params["-L"] or params["--dereference"]:
                     shutil.copytree(paths[0], paths[1], symlinks=True)
@@ -376,18 +410,17 @@ def run_traced(cmd,
                     shutil.copytree(paths[0], paths[1], symlinks=False)
             else:
                 shutil.copy2(paths[0], paths[1])
+            clear_tty(dry_run=dry_run, flush=flush)
             return 0, "", ""
         return sh_any(argv)
 
-    def sh_git(args, verbose=0, dry_run=None):
+    def sh_git(args, verbose=0, dry_run=None, flush=None, humdrum=1):
         if (
             cmd_neutral_if(args, params=['git', 'status'])
             or cmd_neutral_if(args, params=['git', 'branch'], no_switches=['-b'])
             or cmd_neutral_if(args, params=['git', 'remote'], switches=['-v'])
         ):
             dry_run = False
-        # if dry_run:
-        #     return 0, "", ""
         argv, opt_unk, paths, params = simple_parse(
             args, {
                 "-b": "",
@@ -416,41 +449,41 @@ def run_traced(cmd,
                     repo = paths[1].split("/")[-1][: -4]
                     odoo_root = opt_branch
                 if repo:
-                    homes = [os.path.expanduser("~")]
-                    if os.environ.get("TRAVIS_SAVED_HOME_DEVEL"):
+                    homes = [pthuser("~")]
+                    if os_env("TRAVIS_SAVED_HOME_DEVEL"):
                         homes.append(
-                            os.path.dirname(os.environ["TRAVIS_SAVED_HOME_DEVEL"]))
-                    if os.environ.get("HOME_DEVEL"):
+                            pth.dirname(os.environ["TRAVIS_SAVED_HOME_DEVEL"]))
+                    if os_env("HOME_DEVEL"):
                         homes.append(
-                            os.path.dirname(os.environ["HOME_DEVEL"]))
+                            pth.dirname(os.environ["HOME_DEVEL"]))
                     sts, prcout, prcerr = sh_any(
                         ["getent", "passwd", str(os.getuid())])
                     if sts == 0 and prcout:
                         homes.append(prcout.split("\n")[0].split(":")[5])
                     for home in homes:
                         if repo == "tools":
-                            srcpath = os.path.join(home, repo)
+                            srcpath = pth.join(home, repo)
                         elif repo in ("OCB", "odoo"):
-                            srcpath = os.path.join(home, odoo_root)
+                            srcpath = pth.join(home, odoo_root)
                         else:
-                            srcpath = os.path.join(home, odoo_root, repo)
-                        if os.path.isdir(srcpath):
+                            srcpath = pth.join(home, odoo_root, repo)
+                        if pth.isdir(srcpath):
                             break
                         srcpath = ""
         if srcpath:
-            tgt = os.path.join(paths[2] if len(paths) > 2 else "./",
-                               os.path.basename(srcpath))
-            if os.path.isdir(tgt):
+            tgt = pth.join(paths[2] if len(paths) > 2 else "./", pth.basename(srcpath))
+            if pth.isdir(tgt):
                 if verbose:
-                    echo_cmd_verbose(args, dry_run=dry_run)
+                    echo_cmd_verbose(args, dry_run=dry_run, humdrum=humdrum)
+                clear_tty(dry_run=dry_run, flush=flush)
                 return 1, "", ""
             os.mkdir(tgt)
             if repo in ("OCB", "odoo"):
                 for fn in os.listdir(srcpath):
-                    if os.path.isdir(os.path.join(srcpath, fn, ".git")):
+                    if pth.isdir(pth.join(srcpath, fn, ".git")):
                         continue
-                    ffn = os.path.join(srcpath, fn)
-                    if os.path.isdir(ffn):
+                    ffn = pth.join(srcpath, fn)
+                    if pth.isdir(ffn):
                         run_traced(
                             "cp -r %s %s" % (ffn, tgt),
                             verbose=verbose,
@@ -463,7 +496,8 @@ def run_traced(cmd,
                             dry_run=dry_run,
                             is_alias=True)
                 if verbose:
-                    echo_cmd_verbose(args, dry_run=dry_run)
+                    echo_cmd_verbose(args, dry_run=dry_run, humdrum=humdrum)
+                clear_tty(dry_run=dry_run, flush=flush)
                 return 0, "", ""
             else:
                 return run_traced(
@@ -471,27 +505,31 @@ def run_traced(cmd,
                     verbose=verbose,
                     dry_run=dry_run,
                     is_alias=True)
-        return sh_any(argv, verbose=verbose, dry_run=dry_run)
+        return sh_any(
+            argv, verbose=verbose, dry_run=dry_run, flush=flush, humdrum=humdrum)
 
-    def sh_mkdir(args, verbose=0, dry_run=None):
+    def sh_mkdir(args, verbose=0, dry_run=None, flush=None, humdrum=1):
         prcout = prcerr = ""
         if dry_run:
             if verbose:
-                echo_cmd_verbose(args, dry_run=dry_run)
+                echo_cmd_verbose(args, dry_run=dry_run, humdrum=humdrum)
+            clear_tty(dry_run=dry_run, flush=flush)
             return 0, prcout, prcerr
         argv, opt_unk, paths, params = simple_parse(args, {})
         if opt_unk:
             sts, prcout, prcerr = os_system_traced(argv, verbose=verbose)
         else:
             os.mkdir(paths[0])
-            sts = 0 if os.path.exists(paths[0]) else 1
+            sts = 0 if pth.exists(paths[0]) else 1
+        clear_tty(dry_run=dry_run, flush=flush)
         return sts, prcout, prcerr
 
-    def sh_rm(args, verbose=0, dry_run=None):
+    def sh_rm(args, verbose=0, dry_run=None, flush=None, humdrum=1):
         prcout = prcerr = ""
         if dry_run:
             if verbose:
-                echo_cmd_verbose(args, dry_run=dry_run)
+                echo_cmd_verbose(args, dry_run=dry_run, humdrum=humdrum)
+            clear_tty(dry_run=dry_run, flush=flush)
             return 0, prcout, prcerr
         argv, opt_unk, paths, params = simple_parse(
             args,
@@ -501,19 +539,21 @@ def run_traced(cmd,
             }
         )
         tgtpath = paths[0]
-        sts = 0 if os.path.exists(tgtpath) else 1
+        sts = 0 if pth.exists(tgtpath) else 1
         if sts:
             pass
         elif opt_unk or params["-f"]:
-            sts, prcout, prcerr = os_system_traced(argv, verbose=verbose)
+            sts, prcout, prcerr = os_system_traced(
+                argv, verbose=verbose, flush=flush, humdrum=humdrum)
         elif params["-R"]:
             if verbose:
-                echo_cmd_verbose(args, dry_run=dry_run)
+                echo_cmd_verbose(args, flush=flush, humdrum=humdrum)
             shutil.rmtree(tgtpath)
         else:
             if verbose:
-                echo_cmd_verbose(args, dry_run=dry_run)
+                echo_cmd_verbose(args, flush=flush, humdrum=humdrum)
             os.unlink(tgtpath)
+        clear_tty(dry_run=dry_run, flush=flush)
         return sts, prcout, prcerr
 
     args = shlex.split(cmd)
@@ -530,8 +570,461 @@ def run_traced(cmd,
             "cd": sh_cd,
         }.get(args[0])
     if method:
-        return method(args, verbose=verbose, dry_run=dry_run)
-    return sh_any(args, verbose=verbose, dry_run=dry_run)
+        return method(args, verbose=verbose, dry_run=dry_run, humdrum=humdrum)
+    return sh_any(args, verbose=verbose, dry_run=dry_run, flush=flush, humdrum=humdrum)
+
+
+def chain_python_cmd(
+        pyfile,
+        args,
+        verbose=0,
+        dry_run=False,
+        rtime=None,
+        flush=False,
+        humdrum=1):
+    cmd = [sys.executable]
+    fqn = pyfile
+    if "/" not in pyfile:
+        fqn = pth.join(pth.dirname(__file__), pyfile)
+        if not pth.isdir(fqn):
+            fqn = pyfile
+    cmd.append(fqn)
+    for arg in args:
+        cmd.append(arg)
+    rtime = rtime if rtime is not None else (verbose and verbose > 1)
+    return os_system(cmd,
+                     verbose=verbose,
+                     dry_run=dry_run,
+                     rtime=rtime,
+                     flush=flush,
+                     humdrum=humdrum)
+
+
+def chain_python_cmd_traced(
+        pyfile,
+        args,
+        verbose=0,
+        dry_run=False,
+        rtime=None,
+        flush=False,
+        humdrum=1):
+    cmd = [sys.executable]
+    fqn = pyfile
+    if "/" not in pyfile:
+        fqn = pth.join(pth.dirname(__file__), pyfile)
+        if not pth.isdir(fqn):
+            fqn = pyfile
+    cmd.append(fqn)
+    for arg in args:
+        cmd.append(arg)
+    rtime = rtime if rtime is not None else (verbose and verbose > 1)
+    return os_system_traced(cmd,
+                            verbose=verbose,
+                            dry_run=dry_run,
+                            rtime=rtime,
+                            flush=flush,
+                            humdrum=humdrum)
+
+
+class Package(object):
+
+    def __init__(self, path=None):
+        self.name = None
+        self.path = None
+        self.prjname = None
+        self.prjpath = None
+        self.reposname = None
+        self.git_orgid = None
+        self.read_only = False
+        self.url = None
+        self.upstream = None
+        self.stash_list = None
+        self.level = None
+        self.confn = None
+        self.branch = None
+        self.version = None
+        self.release = None
+        self.manifest = None
+        self.testdir = None
+        self.rundir = None
+        self.log_dir = None
+        self.log_fqn = None
+        self.log__daemon_fqn = None
+        self.installable = True
+        self.pyver = ".".join([str(x) for x in sys.version_info[0:2]])
+        self.python_versions = []
+        self.invalid_names = [
+            "build",
+            "debian",
+            "dist",
+            "doc",
+            "docs",
+            "egg-info",
+            "examples",
+            "filestore",
+            "history",
+            "howtos",
+            "images",
+            "junk",
+            "migrations",
+            "openupgrade",
+            "readme",
+            "redhat",
+            "reference",
+            "scripts",
+            "server",
+            "setup",
+            "static",
+            "tests",
+            "tmp",
+            "tools",
+            "travis",
+            # "_travis",
+            "venv_odoo",
+            "win32",
+        ]
+        self.get_info_from_path(path=path)
+
+    def is_odoo_pkg(self, path=None):
+        path = pth.abspath(path or os.getcwd())
+        files = os.listdir(path)
+        filtered = [
+            x
+            for x in files
+            if x in ("__manifest__.py", "__openerp__.py", "__init__.py")
+        ]
+        return len(filtered) == 2 and "__init__.py" in filtered
+
+    def is_pypi_pkg(self, path=None):
+        path = pth.abspath(path or os.getcwd())
+        return (
+            pth.isfile(pth.join(path, "__init__.py")) and (
+                pth.isfile(os.path.join(path, "setup.py"))
+                or pth.isfile(os.path.join(path, "..", "setup.py")))
+        )
+
+    def is_repo_ocb(self, path=None):
+        path = pth.abspath(path or os.getcwd())
+        return (
+            pth.isdir(pth.join(path, ".git"))
+            and (
+                pth.isfile(pth.join(path, "odoo-bin"))
+                or pth.isfile(pth.join(path, "openerp-server"))
+            )
+            and pth.isdir(pth.join(path, "addons"))
+            and (
+                pth.isdir(pth.join(path, "odoo"))
+                and pth.isfile(pth.join(path, "odoo", "__init__.py"))
+            )
+            or (
+                pth.isdir(pth.join(path, "openerp"))
+                and pth.isfile(pth.join(path, "odoo", "__init__.py"))
+            )
+        )
+
+    def is_repo_odoo(self, path=None):
+        path = pth.abspath(path or os.getcwd())
+        if not pth.isdir(pth.join(path, ".git")):
+            return False
+        for fn in os.listdir(path):
+            subpath = pth.join(path, fn)
+            if pth.isdir(subpath) and self.is_odoo_pkg(path=subpath):
+                return True
+        return self.is_repo_ocb(pth.dirname(path))
+
+    def is_repo_pypi(self, path=None):
+        path = pth.abspath(path or os.getcwd())
+        if not pth.isdir(pth.join(path, ".git")):
+            return False
+        for fn in os.listdir(path):
+            subpath = pth.join(path, fn)
+            if pth.isdir(subpath) and pth.isfile(pth.join(subpath, "setup.py")):
+                return True
+        return False
+
+    def get_remote_info(self, path=None):
+        path = pth.abspath(path or os.getcwd())
+        saved_path = pth.abspath(os.getcwd())
+        if path and path != saved_path:
+            os.chdir(path)
+        sts, stdout, stderr = os_system_traced(
+            "git branch", verbose=False, dry_run=False, rtime=False
+        )
+        if sts == 0 and stdout:
+            sts = 123
+            for ln in stdout.split("\n"):
+                if ln.startswith("*"):
+                    self.branch = ln[2:]
+                    sts = 0
+                    break
+            if sts == 0:
+                mo = re.search(r"[0-9]+\.[0-9]+", self.branch)
+                if mo:
+                    self.version = self.release = self.branch[mo.start(): mo.end()]
+        elif self.prjname == "Odoo":
+            mo = re.search(r"[0-9]+\.[0-9]+", path)
+            if not mo:
+                mo = re.search(r"[0-9]+", path)
+            if mo:
+                branch = path[mo.start(): mo.end()]
+                if "." not in branch:
+                    branch += ".0"
+                self.branch = self.version = self.release = branch
+        sts, stdout, stderr = os_system_traced(
+            "git remote -v", verbose=False, dry_run=False, rtime=False)
+        if sts == 0 and stdout:
+            for ln in stdout.split("\n"):
+                if not ln:
+                    continue
+                lns = ln.split()
+                if lns[0] == "origin":
+                    self.url = lns[1]
+                elif lns[0] == "upstream":
+                    self.upstream = lns[1]
+            sts, stdout, stderr = os_system_traced(
+                "git stash list", verbose=False
+            )
+            if sts == 0:
+                self.stash_list = stdout
+        elif self.prjname == "Odoo":
+            self.git_orgid = "oca"
+        elif self.prjname == "pypi":
+            self.git_orgid = "zero"
+        os.chdir(saved_path)
+
+    def data_from_url(self):
+        REV_SHORT_NAMES = {
+            "zeroincombenze": "zero",
+            "OCA": "oca",
+        }
+
+        def get_short_name(uri):
+            git_org = pth.splitext(pth.basename(pth.dirname(uri)))[0]
+            return REV_SHORT_NAMES.get(git_org, git_org)
+
+        if self.url:
+            if "git@github.com:" in self.url:
+                self.git_orgid = get_short_name(self.url.split("@")[1].split(":")[1])
+                self.read_only = False
+            elif "https:" in self.url:
+                self.git_orgid = get_short_name(self.url.split(":")[1])
+                self.read_only = True
+            else:
+                self.git_orgid = get_short_name(self.url)
+                self.read_only = True
+        elif self.git_orgid == "oca":
+            self.read_only = True
+
+    def get_info_from_path(self, path=None):
+        p = pth.abspath(path or os.getcwd())
+        while pth.isdir(p):
+            if pthuser("~").startswith(p) or p == "/":
+                break
+            pkgname = pth.basename(p)
+            if pkgname in self.invalid_names or pkgname.startswith((".", "_")):
+                pass
+            elif self.is_odoo_pkg(path=p):
+                self.path = p
+                self.rundir = p
+                self.name = pkgname
+                self.prjname = "Odoo"
+                self.level = "module"
+                if pth.isfile(pth.join(p, "__manifest__.py")):
+                    self.manifest = pth.join(p, "__manifest__.py")
+                elif pth.isfile(pth.join(p, "__openerp__.py")):
+                    self.manifest = pth.join(p, "__openerp__.py")
+                if pth.isdir(pth.join(p, "tests")):
+                    self.testdir = pth.join(p, "tests")
+            elif self.is_pypi_pkg(path=p):
+                self.rundir = p
+                if not self.name:
+                    self.name = pkgname
+                self.prjname = "pypi"
+                self.level = "module"
+                if pth.isfile(pth.join(p, "setup.py")):
+                    self.manifest = pth.join(p, "setup.py")
+                    self.path = p
+                    self.get_pypi_version()
+                elif pth.isfile(pth.join(pth.dirname(p), "setup.py")):
+                    self.manifest = pth.join(pth.dirname(p), "setup.py")
+                    self.path = pth.dirname(p)
+                    self.get_pypi_version(path=self.path)
+                if pth.isdir(pth.join(p, "tests")):
+                    self.testdir = pth.join(p, "tests")
+            elif self.is_repo_ocb(path=p):
+                if not self.name:
+                    self.name = "OCB"
+                self.prjname = "Odoo"
+                self.prjpath = p
+                if not self.reposname:
+                    self.reposname = "OCB"
+                if not self.level:
+                    self.level = "repo"
+                self.get_odoo_version(path=p)
+                self.get_remote_info(path=p)
+                break
+            elif self.is_repo_odoo(path=p):
+                if not self.name:
+                    self.name = pkgname
+                self.prjname = "Odoo"
+                self.prjpath = p
+                self.reposname = pkgname
+                if not self.level:
+                    self.level = "repo"
+                self.get_remote_info(path=p)
+            elif self.is_repo_pypi(path=p):
+                if not self.name:
+                    self.name = pkgname
+                self.prjname = "pypi"
+                self.prjpath = p
+                # Wrong: just for compatibility with bash
+                self.reposname = "tools"
+                if not self.level:
+                    self.level = "repo"
+                self.get_remote_info(path=p)
+            if self.prjname != "Odoo" and pth.isdir(pth.join(p, ".git")):
+                break
+            p = pth.dirname(p)
+        self.data_from_url()
+        if not self.prjname:
+            self.prjname = os_env("PRJNAME")
+        if not self.reposname:
+            self.reposname = os_env("REPOSNAME")
+        if not self.path:
+            self.path = os_env("PKGPATH", os.getcwd())
+        if not self.git_orgid:
+            self.git_orgid = os_env("GIT_ORGID", "")
+
+    def get_odoo_version(self, path=None):
+        p = pth.abspath(path or os.getcwd())
+        release = pth.join(p, "odoo", "release.py")
+        if not pth.isfile(release):
+            release = pth.join(p, "openerp", "release.py")
+        if os.path.isfile(release):
+            with open(release, RMODE) as fd:
+                for line in fd.read().split("\n"):
+                    mo = re.match(r"version_info *= *\([0-9]+ *, *[0-9]+", line)
+                    if mo:
+                        self.release = "%d.%d" % eval(mo.string.split("=")[1])[0:2]
+                        if not self.version:
+                            self.version = self.release
+                        break
+
+    def get_pypi_version(self, path=None):
+        if os.path.isfile(self.manifest):
+            sts, stdout, stderr = chain_python_cmd_traced(
+                self.manifest, ["--version"])
+            if sts:
+                raise Exception("Wrong manifest file %s" % self.manifest)
+            self.version = stdout.split("\n")[0]
+            if not self.release:
+                self.release = self.version
+
+    def read_manifest_file(self):
+        try:
+            with open(self.manifest, RMODE) as fd:
+                manifest = eval(fd.read())
+        except (ImportError, IOError, SyntaxError):
+            raise Exception("Wrong manifest file %s" % self.manifest)
+        self.version = manifest.get("version", self.release)
+        self.installable = manifest.get("installable", True)
+
+    def get_version_to_log(self):
+        if self.prjname == "Odoo":
+            ver = ("0" + (self.version or "18.0").split(".")[0])[-2:]
+        elif self.prjname == "pypi":
+            ver = (
+                self.pyver.replace(".", "0")
+                if re.match(r"[23]\.[6789]", self.pyver)
+                else self.pyver.replace(".", "")
+            )
+        else:
+            ver = ""
+        return ver
+
+    def get_uniqid(self, ignore_version=False):
+        # UDI (Unique DB Identifier): format "{pkgname}_{git_org}{major_version}"
+        # UMLI (Unique Module Log Identifier):
+        #     format "{git_org}{major_version}.{repos}.{pkgname}"
+        if ignore_version:
+            ver = "..."
+        else:
+            ver = self.get_version_to_log()
+        udi = self.name
+        umli = ""
+        if self.git_orgid and self.git_orgid not in (
+                "oca", "odoo", "zero", "zeroincombenze"):
+            udi += "_" + self.git_orgid
+            umli = self.git_orgid
+        if ver:
+            udi += "_" + ver
+        if self.reposname and self.reposname != "OCB":
+            if umli:
+                umli += "." + self.reposname
+            else:
+                umli = self.reposname
+        if umli:
+            umli += "." + self.name
+        else:
+            umli = self.name
+        if ver and umli:
+            umli += "_" + ver
+        return udi, umli
+
+    def get_log_dir(self, fqn=None):
+        home = pthuser(os_env("TRAVIS_SAVED_HOME") or "~/")
+        if fqn:
+            logdir = pth.dirname(pthuser(pth.abspath(fqn)))
+            self.log_dir = logdir
+        elif self.log_dir:
+            logdir = self.log_dir
+        elif self.read_only:
+            logdir = pth.join(home, "travis_log")
+            self.log_dir = logdir
+        else:
+            testdir = self.testdir
+            if not testdir:
+                testdir = pthuser(pth.join(self.rundir, "tests"))
+            logdir = pth.join(testdir, "logs")
+            self.log_dir = logdir
+        return logdir
+
+    def get_log_filename(self, fqn=None):
+        if fqn:
+            log_fqn = pth.dirname(pthuser(pth.abspath(fqn)))
+            logdir = pth.dirname(log_fqn)
+            log_daemon_fqn = log_fqn.replace(".log", "_nohup.log")
+            self.log_fqn = log_fqn
+            self.log__daemon_fqn = log_daemon_fqn
+        elif self.log_fqn:
+            logdir = self.log_dir
+            log_fqn = self.log_fqn
+            log_daemon_fqn = self.log__daemon_fqn
+        else:
+            logdir = self.get_log_dir()
+            udi, umli = self.get_uniqid()
+            today = datetime.now()
+            today = datetime(today.year, today.month, today.day)
+            suffix = "%s+%05d.log" % (datetime.strftime(today, "%Y%m%d"),
+                                      (datetime.now() - today).seconds)
+            log_fqn = pth.join(logdir, "%s-%s" % (umli, suffix))
+            log_daemon_fqn = pth.join(logdir, "%s_nohup-%s" % (umli, suffix))
+            self.log_fqn = log_fqn
+            self.log__daemon_fqn = log_daemon_fqn
+        return logdir, log_fqn, log_daemon_fqn
+
+    def list_log_filename(self, all_version=True):
+        logdir = self.get_log_dir()
+        udi, umli = self.get_uniqid(ignore_version=all_version)
+        ver = self.get_version_to_log()
+        rex = (
+            r"[0-9]{4}-?[0-9]{2}-?[0-9]{2}(\+[0-9]+)?.log" if not all_version
+            else r"_%s-[0-9]{4}-?[0-9]{2}-?[0-9]{2}(\+[0-9]+)?.log" % ver)
+        log_filenames = []
+        for fn in sorted(os.listdir(logdir), reverse=True):
+            if re.search(rex, fn):
+                log_filenames.append(pth.join(logdir, fn))
+        return log_filenames
 
 
 class CountAction(argparse.Action):
@@ -632,7 +1125,7 @@ class parseoptargs(object):
         i = 1
         valid = False
         while not valid:
-            this_fqn = os.path.abspath(inspect.stack()[i][1])
+            this_fqn = pth.abspath(inspect.stack()[i][1])
             this = nakedname(this_fqn)
             if this in ("__init__", "pdb", "cmd", "z0testlib", "z0lib"):
                 i += 1
@@ -660,16 +1153,16 @@ class parseoptargs(object):
             if isinstance(ODOO_CONF, list):
                 found = False
                 for f in ODOO_CONF:
-                    if os.path.isfile(f):
+                    if pth.isfile(f):
                         ctx['conf_fns'] = (f, ctx['conf_fn'])
                         found = True
                         break
                 if not found:
                     ctx['conf_fns'] = ctx['conf_fn']
             else:
-                if os.path.isfile(ODOO_CONF):
+                if pth.isfile(ODOO_CONF):
                     ctx['conf_fns'] = (ODOO_CONF, ctx['conf_fn'])
-                elif os.path.isfile(OE_CONF):
+                elif pth.isfile(OE_CONF):
                     ctx['conf_fns'] = (OE_CONF, ctx['conf_fn'])
                 else:
                     ctx['conf_fns'] = ctx['conf_fn']
@@ -715,7 +1208,7 @@ class parseoptargs(object):
         ctx = self.create_params_dict(ctx)
         p = 'opt_verbose'
         if p in ctx and ctx[p] is None:
-            if os.environ.get('VERBOSE_MODE', '') in ('0', '1'):
+            if os_env('VERBOSE_MODE', '') in ('0', '1'):
                 ctx[p] = int(os.environ['VERBOSE_MODE'])
             elif os.isatty(0):
                 ctx[p] = 1